backend/pyproject.toml
[project]
name = "data-admin-backend"
version = "0.1.0"
description = "Data Admin backend with Django + DRF"
requires-python = ">=3.12"
dependencies = [
  "asgiref==3.10.0",
  "Django==5.2.8",
  "django-ranged-response==0.2.0",
  "django-simple-captcha==0.6.2",
  "djangorestframework==3.16.1",
  "djangorestframework_simplejwt==5.5.1",
  "drf-spectacular==0.27.1",
  "inflection==0.5.1",
  "jsonschema==4.25.1",
  "jsonschema-specifications==2025.9.1",
  "pillow==12.0.0",
  "PyJWT==2.10.1",
  "PyYAML==6.0.3",
  "referencing==0.37.0",
  "rpds-py==0.29.0",
  "sqlparse==0.5.3",
  "typing_extensions==4.15.0",
  "tzdata==2025.2",
  "uritemplate==4.2.0",
  "cryptography==43.0.1",
  "pymysql>=1.1.2",
  "psycopg2>=2.9.11",
  "trino",
  "pyhive>=0.7.0",
  "openpyxl>=3.1.5",
]

[tool.uv]
dev-dependencies = []

